fix:mmsize_t need support 64bit
VELAPLATFO-3941

+++ mm/mm_heap/mm.h
@@ -145,7 +145,7 @@
 #ifdef CONFIG_MM_SMALL
 typedef uint16_t mmsize_t;
 #else
-typedef uint32_t mmsize_t;
+typedef size_t mmsize_t;
 #endif
 
 /* This describes an allocated chunk.  An allocated chunk is
